+++ application/controllers/info/Fleet.php
@@ -2,32 +2,34 @@
 
 defined('BASEPATH') OR exit('No direct script access allowed');
 
-class Fleet extends Application
+/*
+ * This class acts as an enpoint returning all fleet data in JSON format.
+ * @author Kuanysh
+ */
+class Fleet extends CI_Controller
 {
-
-	public function index()
-	{
+    /**
+    * Fleet endpoint page for the application exposing flights JSON data.
+    *
+    * Maps to the following URL
+    * 		http://comp4711.local/info/flights
+    */
+    public function index() {
         //get all planes in our fleet model
         $fleet = $this->fleetmdl->all();
-        $this->data['pagebody'] = 'fleet';
-        $this->data['fleet'] = $fleet;
-		$this->render();
-	}
-    public function show($planeid)
-    {
-        $plane = $this->fleetmdl->get($planeid);
-        //echo $plane;
-        $this->data['pagebody'] = 'plane';
-        $this->data['id'] = $plane['id'];
-        $this->data['make'] = $plane['make'];
-        $this->data['model'] = $plane['model'];
-        $this->data['price'] = $plane['price'];
-        $this->data['seats'] = $plane['seats'];
-        $this->data['reach'] = $plane['reach'];
-        $this->data['cruise'] = $plane['cruise'];
-        $this->data['takeoff'] = $plane['takeoff'];
-        $this->data['hourly'] = $plane['hourly'];
+        $this->data['pagebody'] = 'info/fleet';
+        $this->data['fleet'] = json_encode($fleet);
         $this->render();
+    }
 
+    /**
+     * Render this page in JSON format
+     */
+    function render()
+    {
+            $this->data['content'] = $this->parser->parse($this->data['pagebody'], $this->data, true);
+            $this->parser->parse('info/fleet', $this->data);
     }
 }
+
+?>