+++ application/views/flights.php
@@ -2,18 +2,36 @@
 <br/><br/>
 
 <table class="table">
-    <thead>
-      <tr>
-          <th>From</th>
-          <th>To</th>
-          <th>Distance</th>
-      </tr>
-    </thead>
+    <tr>
+        <th>
+            Aircraft Code
+        </th>
+        <th>
+            From
+        </th>
+        <th>
+            To
+        </th>
+        <th>
+            Distance
+        </th>
+    </tr>  
     {flights}
-      <tr>
-          <td>{from}</td>
-          <td>{to}</td>
-          <td>{distance}</td>
-      </tr>
+    <tr>
+        <td>
+            <a style="text-decoration: none; color:black;" ref="#" title="The aircraft code is a unique identifier for the aircraft involved in a flight.">
+            {accode}
+            </a>
+        </td>
+        <td>
+            {from}
+        </td>
+        <td>
+            {to}
+        </td>
+        <td>
+            {distance}
+        </td>
+    </tr>
     {/flights}
 </table>